fix generateAuthToken
Moved the generateAuthToken function outside of the app.get handler so that it can be called from multiple places in the code.

Called the generateAuthToken function before the first API request is made, to ensure that the authToken variable has a valid token.

Used the await keyword to wait for the generateAuthToken function to complete before continuing with the API request.